upgrade celery
