+++ care/facility/utils/summarisation/facility_capacity.py
@@ -1,14 +1,5 @@
-from celery.decorators import periodic_task
-from celery.schedules import crontab
 from django.db.models import Sum
-from django.utils.decorators import method_decorator
 from django.utils.timezone import localtime, now
-from django.views.decorators.cache import cache_page
-from django_filters import rest_framework as filters
-from rest_framework import serializers
-from rest_framework.mixins import ListModelMixin
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
-from rest_framework.viewsets import GenericViewSet
 
 from care.facility.api.serializers.facility import FacilitySerializer
 from care.facility.api.serializers.facility_capacity import FacilityCapacitySerializer
@@ -25,64 +16,7 @@ from care.facility.models.inventory import (
 )
 
 
-class FacilitySummarySerializer(serializers.ModelSerializer):
-    facility = FacilitySerializer()
-
-    class Meta:
-        model = FacilityRelatedSummary
-        exclude = (
-            "id",
-            "s_type",
-        )
-
-
-class FacilitySummaryFilter(filters.FilterSet):
-    start_date = filters.DateFilter(field_name="created_date", lookup_expr="gte")
-    end_date = filters.DateFilter(field_name="created_date", lookup_expr="lte")
-    facility = filters.UUIDFilter(field_name="facility__external_id")
-    district = filters.NumberFilter(field_name="facility__district__id")
-    local_body = filters.NumberFilter(field_name="facility__local_body__id")
-    state = filters.NumberFilter(field_name="facility__state__id")
-
-
-class FacilityCapacitySummaryViewSet(
-    ListModelMixin,
-    GenericViewSet,
-):
-    lookup_field = "external_id"
-    queryset = (
-        FacilityRelatedSummary.objects.filter(s_type="FacilityCapacity")
-        .order_by("-created_date")
-        .select_related(
-            "facility",
-            "facility__state",
-            "facility__district",
-            "facility__local_body",
-        )
-    )
-    permission_classes = (IsAuthenticatedOrReadOnly,)
-    serializer_class = FacilitySummarySerializer
-
-    filter_backends = (filters.DjangoFilterBackend,)
-    filterset_class = FacilitySummaryFilter
-
-    @method_decorator(cache_page(60 * 10))
-    def dispatch(self, request, *args, **kwargs):
-        return super().dispatch(request, *args, **kwargs)
-
-    # def get_queryset(self):
-    #     user = self.request.user
-    #     queryset = self.queryset
-    #     if user.is_superuser:
-    #         return queryset
-    #     elif self.request.user.user_type >= User.TYPE_VALUE_MAP["DistrictReadOnlyAdmin"]:
-    #         return queryset.filter(facility__district=user.district)
-    #     elif self.request.user.user_type >= User.TYPE_VALUE_MAP["StateReadOnlyAdmin"]:
-    #         return queryset.filter(facility__state=user.state)
-    #     return queryset.filter(facility__users__id__exact=user.id)
-
-
-def FacilityCapacitySummary():
+def facility_capacity_summary():
     capacity_objects = FacilityCapacity.objects.all().select_related(
         "facility",
         "facility__state",
@@ -140,17 +74,13 @@ def FacilityCapacitySummary():
                 Sum("quantity_in_default_unit")
             )
             if temp1:
-                total_consumed = temp1.get("quantity_in_default_unit__sum", 0)
-                if not total_consumed:
-                    total_consumed = 0
+                total_consumed = temp1.get("quantity_in_default_unit__sum", 0) or 0
             total_added = 0
             temp2 = log_query.filter(is_incoming=True).aggregate(
                 Sum("quantity_in_default_unit")
             )
             if temp2:
-                total_added = temp2.get("quantity_in_default_unit__sum", 0)
-                if not total_added:
-                    total_added = 0
+                total_added = temp2.get("quantity_in_default_unit__sum", 0) or 0
 
             # Calculate Start Stock as
             # end_stock = start_stock - consumption + addition
@@ -207,9 +137,3 @@ def FacilityCapacitySummary():
         facility_summary_obj.save()
 
     return True
-
-
-@periodic_task(run_every=crontab(minute="*/5"))
-def run_midnight():
-    FacilityCapacitySummary()
-    print("Summarised Capacities")
