+++ care/facility/utils/summarisation/patient_summary.py
@@ -1,50 +1,11 @@
-from celery.decorators import periodic_task
-from celery.schedules import crontab
 from django.db.models import Q
-from django.utils.decorators import method_decorator
 from django.utils.timezone import now
-from django.views.decorators.cache import cache_page
-from django_filters import rest_framework as filters
-from rest_framework.mixins import ListModelMixin
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
-from rest_framework.viewsets import GenericViewSet
 
 from care.facility.models import Facility, FacilityRelatedSummary, PatientRegistration
 from care.facility.models.patient_base import BedTypeChoices
-from care.facility.summarisation.facility_capacity import (
-    FacilitySummaryFilter,
-    FacilitySummarySerializer,
-)
 
 
-class PatientSummaryViewSet(ListModelMixin, GenericViewSet):
-    lookup_field = "external_id"
-    queryset = FacilityRelatedSummary.objects.filter(s_type="PatientSummary").order_by(
-        "-created_date"
-    )
-    permission_classes = (IsAuthenticatedOrReadOnly,)
-    serializer_class = FacilitySummarySerializer
-
-    filter_backends = (filters.DjangoFilterBackend,)
-    filterset_class = FacilitySummaryFilter
-
-    @method_decorator(cache_page(60 * 10))
-    def dispatch(self, request, *args, **kwargs):
-        return super().dispatch(request, *args, **kwargs)
-
-    # def get_queryset(self):
-    #     user = self.request.user
-    #     queryset = self.queryset
-    #     if user.is_superuser:
-    #         return queryset
-    #     elif self.request.user.user_type >= User.TYPE_VALUE_MAP["DistrictReadOnlyAdmin"]:
-    #         return queryset.filter(facility__district=user.district)
-    #     elif self.request.user.user_type >= User.TYPE_VALUE_MAP["StateReadOnlyAdmin"]:
-    #         return queryset.filter(facility__state=user.state)
-    #     return queryset.filter(facility__users__id__exact=user.id)
-
-
-def PatientSummary():
+def patient_summary():
     facility_objects = Facility.objects.all()
     patient_summary = {}
     for facility_object in facility_objects:
@@ -120,9 +81,7 @@ def PatientSummary():
             )
             facility.created_date = now()
             facility.data.pop("modified_date")
-            if facility.data == patient_summary[i]:
-                pass
-            else:
+            if facility.data != patient_summary[i]:
                 facility.data = patient_summary[i]
                 latest_modification_date = now()
                 facility.data.update(
@@ -142,9 +101,3 @@ def PatientSummary():
                 s_type="PatientSummary", facility_id=i, data=patient_summary[i]
             ).save()
     return True
-
-
-@periodic_task(run_every=crontab(hour="*/1", minute=59))
-def run_midnight():
-    PatientSummary()
-    print("Summarised Patients")
