+++ care/facility/utils/summarisation/tests_summary.py
@@ -1,46 +1,7 @@
-from celery.decorators import periodic_task
-from celery.schedules import crontab
 from django.core.exceptions import ObjectDoesNotExist
 from django.utils import timezone
-from django.utils.decorators import method_decorator
-from django.views.decorators.cache import cache_page
-from django_filters import rest_framework as filters
-from rest_framework.mixins import ListModelMixin
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
-from rest_framework.viewsets import GenericViewSet
 
 from care.facility.models import Facility, FacilityRelatedSummary, PatientSample
-from care.facility.summarisation.facility_capacity import (
-    FacilitySummaryFilter,
-    FacilitySummarySerializer,
-)
-
-
-class TestsSummaryViewSet(ListModelMixin, GenericViewSet):
-    lookup_field = "external_id"
-    queryset = FacilityRelatedSummary.objects.filter(s_type="TestSummary").order_by(
-        "-created_date"
-    )
-    permission_classes = (IsAuthenticatedOrReadOnly,)
-    serializer_class = FacilitySummarySerializer
-
-    filter_backends = (filters.DjangoFilterBackend,)
-    filterset_class = FacilitySummaryFilter
-
-    # def get_queryset(self):
-    #     user = self.request.user
-    #     queryset = self.queryset
-    #     if user.is_superuser:
-    #         return queryset
-    #     elif self.request.user.user_type >= User.TYPE_VALUE_MAP["DistrictReadOnlyAdmin"]:
-    #         return queryset.filter(facility__district=user.district)
-    #     elif self.request.user.user_type >= User.TYPE_VALUE_MAP["StateReadOnlyAdmin"]:
-    #         return queryset.filter(facility__state=user.state)
-    #     return queryset.filter(facility__users__id__exact=user.id)
-
-    @method_decorator(cache_page(60 * 60 * 10))
-    def dispatch(self, request, *args, **kwargs):
-        return super().dispatch(request, *args, **kwargs)
 
 
 def tests_summary():
@@ -84,29 +45,20 @@ def tests_summary():
             )
             facility_test_summary.created_date = timezone.now()
             facility_test_summary.data.pop("modified_date")
-            if not facility_test_summary.data == facility_tests_summarised_data:
+            if facility_test_summary.data != facility_tests_summarised_data:
                 facility_test_summary.data = facility_tests_summarised_data
                 latest_modification_date = timezone.now()
-                facility_test_summary.data.update(
-                    {
-                        "modified_date": latest_modification_date.strftime(
-                            "%d-%m-%Y %H:%M"
-                        )
-                    }
-                )
+                facility_test_summary.data[
+                    "modified_date"
+                ] = latest_modification_date.strftime("%d-%m-%Y %H:%M")
                 facility_test_summary.save()
         except ObjectDoesNotExist:
             modified_date = timezone.now()
-            facility_tests_summarised_data.update(
-                {"modified_date": modified_date.strftime("%d-%m-%Y %H:%M")}
+            facility_tests_summarised_data["modified_date"] = modified_date.strftime(
+                "%d-%m-%Y %H:%M"
             )
             FacilityRelatedSummary.objects.create(
                 s_type="TestSummary",
                 facility=facility,
                 data=facility_tests_summarised_data,
             )
-
-
-@periodic_task(run_every=crontab(hour=23, minute=59))
-def run_midnight():
-    tests_summary()
