+++ care/utils/notification_handler.py
@@ -1,6 +1,6 @@
 import json
 
-import celery
+from celery import shared_task
 from django.apps import apps
 from django.conf import settings
 from pywebpush import WebPushException, webpush
@@ -23,12 +23,12 @@ class NotificationCreationException(Exception):
     pass
 
 
-@celery.task()
+@shared_task
 def notification_task_generator(**kwargs):
     NotificationGenerator(**kwargs).generate()
 
 
-@celery.task()
+@shared_task
 def send_webpush(**kwargs):
     user = User.objects.get(username=kwargs.get("username"))
     message = kwargs.get("message")
