+++ requirements/base.txt
@@ -43,7 +43,7 @@ littletable==2.0.7
 
 # Asyncronous Part
 # ------------------------------------------------------------------------------
-celery==4.4.2
+celery==5.3.0 # https://docs.celeryq.dev/en/latest/changelog.html
 
 # WebPushes
 # ------------------------------------------------------------------------------
